hw4/requirements.txt
Cython==0.25.2
matplotlib==2.0.2
nltk==3.2.2
numpy==1.12.1
Pillow==2.3.0
scikit-image >= 0.13.0
scikit-learn==0.18.1
scipy==0.19.0
word2vec==0.9.1
adjustText >= 0.6.1

